Try to avoid using up all compute cycles
When running this script every minute, I got this error a few times:

  Service invoked too many times in a short time: gmail rateMax.

And then I eventually got this error:

  Service invoked too many times for one day: gmail.

After getting those errors, I applied this change to the script, and
it's been running with this change in place for a week. I haven't seen
those errors since applying this change, so let's hope that this issue
is resolved.